+++ src/txio.C
@@ -414,7 +414,7 @@ auto io_ns::get_enc_param(long enc, long pos) -> long {
     return to_signed(custom_table[to_unsigned(enc)][to_unsigned(pos)].value);
 }
 
-void LinePtr::change_encoding(int wc) {
+void LinePtr::change_encoding(long wc) {
     if (wc >= 0 && wc < to_signed(max_encoding)) {
         cur_encoding = to_unsigned(wc);
         the_log << lg_start_io << "Input encoding changed to " << wc << " for " << file_name << lg_end;
